Got the updater working

+++ ConvenienceShare/src/org/cnv/shr/dmn/mn/GenerateKey.java
@@ -1,6 +1,7 @@
 
 package org.cnv.shr.dmn.mn;
 
+import java.net.InetAddress;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -8,6 +9,7 @@ import java.nio.file.Paths;
 import org.cnv.shr.dmn.Services;
 import org.cnv.shr.dmn.UpdateManager;
 import org.cnv.shr.stng.Settings;
+import org.cnv.shr.updt.UpdateInfo;
 import org.cnv.shr.updt.UpdateInfoImpl;
 import org.cnv.shr.util.KeysService;
 
@@ -17,10 +19,8 @@ public class GenerateKey
 	
 	public static void main(String[] args) throws Exception
 	{
-		// Junk...
-		String ip = "127.0.0.1";
-		int port = 7005;
-		
+		String ip = InetAddress.getLocalHost().getHostAddress();
+		int port = UpdateInfo.DEFAULT_UPDATE_PORT;
 
 		// Generate a new key for the updater
 		Path keysFile = root.resolve(UpdateInfoImpl.KEYS_TXT);
@@ -35,6 +35,11 @@ public class GenerateKey
 		Services.settings.codeUpdateKey.set(root.resolve("updateKey"));
 		updateManager.updateInfo(ip, port, service.getPublicKey());
 		updateManager.write();
-		Files.delete(settingsFile);
+		if (Files.exists(settingsFile))
+		{
+			Files.delete(settingsFile);
+		}
+		
+		System.out.println("Done.");
 	}
 }
